pages/requirements.txt
altair==5.5.0
beautifulsoup4==4.12.3
common==0.1.2
matplotlib==3.9.3
mpld3==0.5.10
numpy==2.2.0
pandas==2.2.3
Pillow==11.0.0
plotly==5.24.1
Requests==2.32.3
streamlit==1.40.1
